CI/python-coroutine/download_server/requirements.txt
prettytable
requests
aiohttp
aiofiles
scipy
pillow
numpy
